Sudipto - Task ID #407#,#445# -2017-03-14 6:40pm
Task ID #407 is done
Task ID #445 is partially done. As I dont have the access to the three sites so I could not test it there. Source and Employee Type is getting entered in database but its not showing in staff section of the aspx page. Please fix it in http://web.jmgrovebuildingsupply.com/ViewApplicantUser.aspx page.